Use fab-classic instead of fabric.
fab-classic is a package which adds Python3 support
for Fabric 1.x. The official fabric package supports Python3
but it adds support only in version 2.x. Version 2.x fabric
has a completely diffrent API and we need to rewrite everything
from scratch if we want to go with the version 2.x. So we
switch to fab-classic to run this on Python3.