setup: Remove Erlang magic cookie, touch rabbitmq config.
Changing the `rabbitmq-env.conf` file is what triggers the
re-initialization of rabbitmq users during setup -- which is necessary
because the `zulip` user's password in rabbitmq has been
regenerated.

`configure-rabbitmq` will also generate a new Erlang magic cookie for
rabbitmq, but only if it is missing or from the bad randomizer.  Since
the Erlang cookie is effectively a secret (though, as of Zulip Server
4.9, properly protected by only listening on localhost), we remove it
and allow `configure-rabbitmq` to regenerate it.

Since any Erlang process will regenerate the cookie if it is missing,
this does cause a window, after boot and before configuration, where
RabbitMQ has an Erlang magic cookie generated by the bad Erlang
randomizer.  This is not a risk because the rabbitmq distribution port
only ever binds to localhost.