Disable apt-daily at the beginning of the build
This should increase reliability of the image build job - we tend to
have non-deterministic failures due to apt not being able to obtain the
lock. By disabling these services that can run apt in the background at
the beginning instead of only in 01-initial-setup, we should run into
issues less often.